backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
google-generativeai==0.8.3

